Refactor data retrieval and updating methods in DataController
This commit changes the way data is retrieved and updated from the database. The getDataById and getDataBySupervisor methods retrieve additional related data for the response. The edit method is removed and some of its functionality is incorporated into the updateSchool and update methods. The searchSchoolFilter method is refactored to simplify the querying.